GET - /api/posts/top/{category} : ファンサイトTOPの「企画情報」「寄贈情報」「その他情報」用(それぞれのカテゴリで新着4件)

+++ src/main/java/com/example/zoutohanafansitedemo/entity/post/Post.java
@@ -6,13 +6,13 @@ import java.time.LocalDateTime;
 
 @Data
 public class Post {
-    private Integer id;
-    private Integer adminId;
-    private Integer category;
+    private long id;
+    private int adminId;
+    private int category;
     private String title;
     private String content;
     private LocalDateTime postedAt;
-    private Integer status;
+    private int status;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
